fix: remove periodic docker system prune (#80)
The periodic system docker prune would clean up also dangling images and this may cause trouble if a container is only run once at startup and then exits, its image is dangling and would be removed. On the next restart, if no internet connection is available, the image cannot be loaded and the container cannot start.